Fix CS0266 that left main unable to compile
CurrentPlayerSendCooldownTicks subtracts two SimulationTick.Value longs
and returned the result directly from an int method, so Assembly-CSharp
failed to build - a clean batch compile of origin/main aborts with
"Scripts have compiler errors". Adds the explicit narrowing cast;
remaining is bounded by the 30-tick cooldown so it cannot overflow.

Caught because a batch Unity run in a separate worktree could not
compile. The commit that introduced it noted it was verified against a
running editor's log rather than a batch compile, which is how a hard
compile error got through.

+++ unity/LTW.UnityClient/Assets/Scripts/Simulation/UnityCommandAdapter.cs
@@ -45,8 +45,11 @@ namespace LTW.UnityClient.Simulation
             }
 
             var snapshot = simulation.GetSnapshot();
+            // SimulationTick.Value is a long, so this subtraction is a long and needs an explicit
+            // narrowing cast to match this method's int return (CS0266 without it — main did not
+            // compile). Safe: remaining is bounded above by the send cooldown itself (30 ticks).
             var remaining = snapshot.Players.Get(simulation.LocalPlayerId).NextSendAvailableTick.Value - snapshot.Tick.Value;
-            return remaining > 0 ? remaining : 0;
+            return remaining > 0 ? (int)remaining : 0;
         }
 
         public VerticalSliceCommandResult PreviewSampleTower(int x, int y) => PreviewTower(SampleVerticalSliceContent.TowerId, x, y);
